move files to correct directory
