+++ packages/dds/tree/src/test/sequence-change-family/randomSequenceGenerator.ts
@@ -4,10 +4,10 @@
  */
 
 import { makeRandom } from "@fluid-internal/stochastic-test-utils";
-import { AnchorSet, FieldKey, UpPath } from "../tree";
-import { SequenceEditBuilder, singleTextCursor } from "../feature-libraries";
-import { jsonString } from "../domains";
-import { Transposed as T } from "../changeset";
+import { AnchorSet, FieldKey, UpPath } from "../../tree";
+import { SequenceEditBuilder, singleTextCursor } from "../../feature-libraries";
+import { jsonString } from "../../domains";
+import { Transposed as T } from "../../changeset";
 
 /**
  *
