+++ packages/dds/tree/src/tree/index.ts
@@ -21,7 +21,6 @@ export * from "./pathTree";
 export * from "./anchorSet";
 export * from "./treeTextFormat";
 export * from "./visitDelta";
-export * from "./fuzz";
 
 // Split this up into separate import and export for compatibility with API-Extractor.
 import * as Delta from "./delta";
