+++ packages/dds/tree/src/util/index.ts
@@ -17,4 +17,3 @@ export * from "./typeCheck";
 export * from "./brand";
 export * from "./offsetList";
 export * from "./stackyIterator";
-export * from "./randomSequenceGenerator";
